Add string filter function and swap integers function
In this commit, two functions were included. Firstly, a new function `getStringsOnly` was added to `Main.kt` that filters a list for string type elements. This was done to refine the list processing capability of the application. Secondly, a new file `swap.kt` was created containing a function `swapInts` which swaps the positions of two integers in an array, intended to increase the application's ability to manipulate arrays.