Add new Callable task in 'callable3.kt' file
A new Kotlin file called 'callable3.kt' is added in the 'threads/callable' package. This file contains a main method which creates an executor service to handle Callable tasks. The tasks sleep for a certain amount of time, output relevant information including the thread ID then return a predefined value.